Improve physics feel and behavior <bug

+++ core/src/main/java/com/group5/core/world/WorldManager.java
@@ -8,6 +8,7 @@ import com.badlogic.gdx.physics.box2d.ContactImpulse;
 import com.badlogic.gdx.physics.box2d.ContactListener;
 import com.badlogic.gdx.physics.box2d.Manifold;
 import com.badlogic.gdx.physics.box2d.World;
+import com.badlogic.gdx.physics.box2d.WorldManifold;
 import com.group5.core.controllers.Spawner;
 
 import java.util.ArrayList;
@@ -18,6 +19,12 @@ import java.util.List;
  * Functions as a container for spawned items in the game.
  */
 public class WorldManager {
+
+    /**
+     * The scale with which pixel values have to be multiplied to get world-scale values.
+     */
+    public static final float PHYSICS_SCALE_FACTOR = 1 / 50.f;
+
     /**
      * The Box2D physics world.
      */
@@ -57,7 +64,7 @@ public class WorldManager {
      * Constructs a new, empty world with a default gravity.
      */
     public WorldManager() {
-        physicsWorld = new World(new Vector2(0, -20), true);
+        physicsWorld = new World(new Vector2(0, -50), true);
 
         objects = new ArrayList<WorldObject>();
         spawner = new Spawner(this);
@@ -235,7 +242,7 @@ public class WorldManager {
         while (wIter.hasNext()) {
             w = wIter.next();
 
-            if ((w.getX() * 50.f + w.getWidth() * 50.f) < (player.getX() * 50.f - 1000)) {
+            if ((w.getX() + w.getWidth()) < (player.getX() - 20)) {
                 if (w.getPhysicsBody() != null) {
                     physicsWorld.destroyBody(w.getPhysicsBody());
                 }
@@ -287,9 +294,17 @@ public class WorldManager {
 
         @Override
         public void beginContact(final Contact contact) {
-            if (contact.getFixtureA().getUserData() == player && contact.getFixtureB().getUserData() instanceof Obstacle
-                    || contact.getFixtureB().getUserData() == player && contact.getFixtureA().getUserData() instanceof Obstacle) {
-                player.kill();
+            if (contact.getFixtureA().getUserData() == player
+                    && contact.getFixtureB().getUserData() instanceof Obstacle
+                    || contact.getFixtureB().getUserData() == player
+                    && contact.getFixtureA().getUserData() instanceof Obstacle) {
+                WorldManifold w = contact.getWorldManifold();
+                float angle = w.getNormal().angle();
+                if (angle < 45
+                        || angle > 135 && angle < 225
+                        || angle > 315) {
+                    player.kill();
+                }
             }
         }
